+++ perception/obstacle_segmentation/occupancy_grid_map_outlier_filter/package.xml
@@ -1,8 +1,8 @@
 <?xml version="1.0"?>
 <package format="3">
-  <name>outlier_filter</name>
+  <name>occupancy_grid_map_outlier_filter</name>
   <version>0.1.0</version>
-  <description>The ROS2 outlier_filter package</description>
+  <description>The ROS2 occupancy_grid_map_outlier_filter package</description>
 
   <maintainer email="abrahammonrroy@yahoo.com">amc-nu</maintainer>
   <maintainer email="yukihiro.saito@tier4.jp">Yukihiro Saito</maintainer>
